Data.LCA.Online

Description

Provides online calculation of the the lowest common ancestor in O(log h) by compressing the spine of a Path using a skew-binary random access list.

This library implements the technique described in my talk

http://www.slideshare.net/ekmett/skewbinary-online-lowest-common-ancestor-search

to improve the known asymptotic bounds on online lowest common ancestor search.

Algorithms used here assume that the key values chosen for k are globally unique.

Documentation

data Path a

Compressed paths using skew binary random access lists

lca :: Path a -> Path b -> Path a

O(log h) Compute the lowest common ancestor of two paths.

empty :: Path a

The empty Path

cons :: Int -> a -> Path a -> Path a

O(1) Invariant: most operations assume that the keys k are globally unique

Extend the Path with a new node ID and value.

uncons :: Path a -> Maybe (Int, a, Path a)

O(1) Extract the node ID and value from the newest node on the Path.

view :: Path a -> View Path a

O(1) Extract the node ID and value from the newest node on the Path, slightly faster than uncons.

null :: Path a -> Bool

O(1) Returns True iff the path is empty.

length :: Path a -> Int

O(1) Determine the length of a Path.

isAncestorOf :: Path a -> Path b -> Bool

O(log h) xs isAncestorOf ys holds when xs is a prefix starting at the root of Path ys.

keep :: Int -> Path a -> Path a

O(log (h - k)) to keep k elements of Path of length h

drop :: Int -> Path a -> Path a

O(log k) to drop k elements from a Path

traverseWithKey :: Applicative f => (Int -> a -> f b) -> Path a -> f (Path b)

Traverse a Path with access to the node IDs.

toList :: Path a -> [(Int, a)]

Convert a Path to a list of (ID, value) pairs.

fromList :: [(Int, a)] -> Path a

Build a Path from a list of (ID, value) pairs.

(~=) :: Path a -> Path b -> Bool

O(1) Compare to see if two trees have the same leaf key